download: install_data: add keep argument

+++ spinalcordtoolbox/download.py
@@ -104,12 +104,13 @@ def unzip(compressed, dest_folder):
         raise
 
 
-def install_data(url, dest_folder):
+def install_data(url, dest_folder, keep=False):
     """
     Download a data bundle from a URL and install in the destination folder.
 
     :param url: URL or sequence thereof (if mirrors).
-    :param dest_folder: destination directory. Will be cleaned if exists.
+    :param dest_folder: destination directory. Will be cleaned if exists,
+                        unless `keep` is True.
     :return: None
 
     Notes:
@@ -120,10 +121,11 @@ def install_data(url, dest_folder):
       from files that would be in the bundle (not the rest).
     """
 
-    if os.path.exists(dest_folder):
+    if not keep and os.path.exists(dest_folder):
         shutil.rmtree(dest_folder)
-
-    os.makedirs(dest_folder)
+        os.makedirs(dest_folder)
+    else:
+        os.makedirs(dest_folder, exist_ok=True)
 
     tmp_file = download_data(url)
 
@@ -161,14 +163,23 @@ def install_data(url, dest_folder):
             srcpath = os.path.join(cwd, d)
             relpath = os.path.relpath(srcpath, bundle_folder)
             dstpath = os.path.join(dest_folder, relpath)
-            logger.debug("- d+ %s", relpath)
-            os.makedirs(dstpath)
+            if os.path.exists(dstpath):
+                # lazy -- we assume it's a directory, otherwise it will crash safely
+                logger.debug("- d- %s", relpath)
+            else:
+                logger.debug("- d+ %s", relpath)
+                os.makedirs(dstpath)
 
         for f in fs:
             srcpath = os.path.join(cwd, f)
             relpath = os.path.relpath(srcpath, bundle_folder)
             dstpath = os.path.join(dest_folder, relpath)
-            logger.debug("- f+ %s", relpath)
+            if os.path.exists(dstpath):
+                logger.debug("- f! %s", relpath)
+                logger.warning("Updating existing %s", dstpath)
+                os.unlink(dstpath)
+            else:
+                logger.debug("- f+ %s", relpath)
             shutil.copy(srcpath, dstpath)
 
     logger.info("Removing temporary folders...")
